Fix Live Scoreboard Load
The communication needed timeouts since the page loads faster than the database can respond.

Also needed to fix how the two pages were communicating.

There needs to be some sort of safeguard here checking if live scoreboard is populated (idk if it'll throw an error if it's not).

And Im also unsure if a trigger from the db will update correctly or add duplicate rows. It shouldn't since they use the same function.

This fixed the bug but a lot more needs to be cleaned up in the cleanup card essentially.